requirements.txt
django>=3.2,<4.0
djongo>=1.3.6
djangorestframework>=3.12.4
pymongo>=3.11.4
aiomysql